refactored ord2.context / core.context modules; absorbed core.viewctx
