+++ ordec/core/ordb.py
@@ -937,7 +937,7 @@ class Node(tuple, metaclass=NodeMeta, build_node=False):
 
     def ctx(self):
         """Return a Context for use as a context manager: ``with node.ctx(): ...``"""
-        from ordec.ord2.context import NodeContext
+        from .context import NodeContext
         return NodeContext(self)
 
     def __copy__(self) -> 'Self':
